Make tests pass on windows

+++ src/test/scala/com/microsoft/ml/spark/downloader/DownloaderSuite.scala
@@ -69,7 +69,9 @@ class DownloaderSuite extends TestBase {
   }
 
   override def afterAll(): Unit = {
-    FileUtils.forceDelete(saveDir)
+    if (saveDir.exists()) {
+      FileUtils.forceDelete(saveDir)
+    }
     super.afterAll()
   }
